symbolic link check

+++ install.sh
@@ -9,7 +9,7 @@ do
     [ -x $DIR/setup.sh ] && $DIR/setup.sh
     for FILE in $DIR/.??*
     do
-        ln -Fis "$PWD/$FILE" $HOME
+        [ -f $PWD/$FILE ] && ln -Fis "$PWD/$FILE" $HOME
     done
 done
